Nest perihelion's `Frame` inside `Websocket`
A WebSocket frame is only meaningful as part of the protocol, and the name
was colliding with ultimatum's layout frame in the umbrella, where nothing
at the use site showed which one `import soundness.*` had supplied.
Ultimatum's is now the only top-level `Frame`.

`unmask` had to change to move: reading an element of a `Data`
(`Array[Byte]^{}`) inside the enclosing object is required at `^{any}`,
which a read-only capture set cannot subsume — the same inference that has
blocked `TarHeader`, `HpackTable` and `MathmlReader`. Here `readUnchecked`
is enough, which those three are worth re-testing against.
